@@ -1,8 +1,8 @@
 package uk.joshiejack.husbandry.api;
 
-import net.minecraft.entity.MobEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import uk.joshiejack.penguinlib.util.icon.Icon;
 
 import javax.annotation.Nullable;
@@ -13,5 +13,5 @@ public interface IProducts {
 
     int getDaysBetweenProducts();
 
-    List<ItemStack> getProduct(MobEntity entity, @Nullable PlayerEntity player);
+    List<ItemStack> getProduct(Mob entity, @Nullable Player player);
 }
