@@ -1,9 +1,10 @@
 package uk.joshiejack.husbandry.api.trait;
 
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.Hand;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.player.Player;
 import uk.joshiejack.husbandry.api.IMobStats;
 
 public interface IInteractiveTrait extends IMobTrait {
-    boolean onRightClick(IMobStats<?> stats, PlayerEntity player, Hand hand);
+    boolean onEntityInteract(Mob mob, IMobStats<?> stats, Player player, InteractionHand hand);
 }
