@@ -1,7 +1,8 @@
 package uk.joshiejack.husbandry.api.trait;
 
+import net.minecraft.world.entity.Mob;
 import uk.joshiejack.husbandry.api.IMobStats;
 
 public interface IJoinWorldTrait extends IMobTrait {
-    void onJoinWorld(IMobStats<?> stats);
+    void onJoinWorld(Mob mob, IMobStats<?> stats);
 }
