@@ -1,7 +1,8 @@
 package uk.joshiejack.husbandry.api.trait;
 
+import net.minecraft.world.entity.Mob;
 import uk.joshiejack.husbandry.api.IMobStats;
 
 public interface INewDayTrait extends IMobTrait {
-    void onNewDay(IMobStats<?> stats);
+    void onNewDay(Mob mob, IMobStats<?> stats);
 }
