@@ -1,39 +1,31 @@
 package uk.joshiejack.husbandry.client;
 
-import net.minecraft.client.gui.ScreenManager;
-import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.RenderTypeLookup;
-import net.minecraft.client.renderer.model.IBakedModel;
-import net.minecraft.client.renderer.model.ModelResourceLocation;
-import net.minecraft.entity.player.PlayerInventory;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.client.event.ModelBakeEvent;
-import net.minecraftforge.client.event.ModelRegistryEvent;
-import net.minecraftforge.client.event.TextureStitchEvent;
-import net.minecraftforge.client.model.ModelLoader;
-import net.minecraftforge.client.model.data.ModelProperty;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import net.minecraft.client.resources.model.BakedModel;
+import net.minecraft.client.resources.model.ModelResourceLocation;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.api.distmarker.OnlyIn;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.neoforge.client.event.ModelEvent;
+import net.neoforged.neoforge.client.event.RegisterMenuScreensEvent;
+import net.neoforged.neoforge.client.model.data.ModelProperty;
 import uk.joshiejack.husbandry.Husbandry;
-import uk.joshiejack.husbandry.block.HusbandryBlocks;
 import uk.joshiejack.husbandry.client.gui.PageStats;
 import uk.joshiejack.husbandry.client.renderer.EggSupplierBakedModel;
-import uk.joshiejack.husbandry.tileentity.IncubatorTileEntity;
-import uk.joshiejack.husbandry.tileentity.NestTileEntity;
+import uk.joshiejack.husbandry.world.block.entity.IncubatorBlockEntity;
+import uk.joshiejack.husbandry.world.block.entity.NestTileEntity;
 import uk.joshiejack.penguinlib.client.gui.book.Book;
 import uk.joshiejack.penguinlib.client.gui.book.tab.NotesTab;
 import uk.joshiejack.penguinlib.client.gui.book.tab.Tab;
-import uk.joshiejack.penguinlib.inventory.AbstractBookContainer;
 import uk.joshiejack.penguinlib.util.icon.Icon;
 import uk.joshiejack.penguinlib.util.icon.TextureIcon;
+import uk.joshiejack.penguinlib.world.inventory.AbstractBookMenu;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -45,60 +37,63 @@ public class HusbandryClient {
     public static final ResourceLocation FODDER = new ResourceLocation(Husbandry.MODID, "block/fodder");
     public static final ResourceLocation SLOP = new ResourceLocation(Husbandry.MODID, "block/slop");
 
-    @SubscribeEvent
-    public static void onMapping(TextureStitchEvent.Pre event) {
-        event.addSprite(FODDER);
-        event.addSprite(SLOP);
-    }
+    //TODO: DATA GEN THESE?
+//    @SubscribeEvent
+//    public static void onMapping(TextureStitchEvent.Pre event) {
+//        event.addSprite(FODDER);
+//        event.addSprite(SLOP);
+//    }
 
     @SubscribeEvent
-    public static void setup(FMLClientSetupEvent event) {
-        ScreenManager.register(Husbandry.HusbandryContainers.BOOK.get(), ((AbstractBookContainer container, PlayerInventory inv, ITextComponent text) ->
+    public static void setup(RegisterMenuScreensEvent event) {
+        event.register(Husbandry.HusbandryContainers.BOOK.get(), ((AbstractBookMenu container, Inventory inv, Component text) ->
                 Book.getInstance(Husbandry.MODID, container, inv, text, (Book bs) -> {
-                    ITextComponent stats = new TranslationTextComponent("gui." + Husbandry.MODID + ".stats");
-                    ITextComponent notes = new TranslationTextComponent("gui." + Husbandry.MODID + ".notes");
+                    Component stats = Component.translatable("gui." + Husbandry.MODID + ".stats");
+                    Component notes = Component.translatable("gui." + Husbandry.MODID + ".notes");
                     bs.withTab(new Tab(stats, PageStats.ICON)).withPage(new PageStats(stats));
                     bs.withTab(new NotesTab(notes, new TextureIcon(Icon.DEFAULT_LOCATION, 0, 0))
                             .withCategory(new ResourceLocation(Husbandry.MODID, "care_category")));
                 })
         ));
 
-        RenderType cutout = RenderType.cutout();
-        RenderTypeLookup.setRenderLayer(HusbandryBlocks.FEEDING_TRAY.get(), cutout);
-        RenderTypeLookup.setRenderLayer(HusbandryBlocks.NEST.get(), cutout);
-        RenderTypeLookup.setRenderLayer(HusbandryBlocks.BOWL.get(), cutout);
-        RenderTypeLookup.setRenderLayer(HusbandryBlocks.INCUBATOR.get(), cutout);
-        RenderTypeLookup.setRenderLayer(HusbandryBlocks.TRUFFLE_BLOCK.get(), cutout);
-        RenderTypeLookup.setRenderLayer(HusbandryBlocks.TROUGH.get(), cutout);
+        //TODO: ADD TO THE MODEL FILES
+//        RenderType cutout = RenderType.cutout();
+//        RenderTypeLookup.setRenderLayer(HusbandryBlocks.FEEDING_TRAY.get(), cutout);
+//        RenderTypeLookup.setRenderLayer(HusbandryBlocks.NEST.get(), cutout);
+//        RenderTypeLookup.setRenderLayer(HusbandryBlocks.BOWL.get(), cutout);
+//        RenderTypeLookup.setRenderLayer(HusbandryBlocks.INCUBATOR.get(), cutout);
+//        RenderTypeLookup.setRenderLayer(HusbandryBlocks.TRUFFLE_BLOCK.get(), cutout);
+//        RenderTypeLookup.setRenderLayer(HusbandryBlocks.TROUGH.get(), cutout);
     }
 
     public static final ResourceLocation DEFAULT_NEST_EGG = new ResourceLocation(Husbandry.MODID, "block/nest_egg");
     public static final ResourceLocation DEFAULT_INCUBATOR_EGG = new ResourceLocation(Husbandry.MODID, "block/incubator_egg");
 
     @SubscribeEvent
-    public static void loadModels(ModelRegistryEvent event) {
-        ModelLoader.addSpecialModel(DEFAULT_NEST_EGG);
-        ModelLoader.addSpecialModel(DEFAULT_INCUBATOR_EGG);
+    public static void loadModels(ModelEvent.RegisterAdditional event) {
+        event.register(DEFAULT_NEST_EGG);
+        event.register(DEFAULT_INCUBATOR_EGG);
     }
 
-    private static void replaceWithEggSupplier(String model, ModelBakeEvent event, Map<Item, ResourceLocation> overrides, ModelProperty<ItemStack> property) {
-        ResourceLocation key = new ModelResourceLocation(model);
-        IBakedModel nest = event.getModelManager().getModel(key);
-        if (!(event.getModelRegistry().get(key) instanceof EggSupplierBakedModel))
-            event.getModelRegistry().put(key, new EggSupplierBakedModel(nest, overrides, property));
+    private static void replaceWithEggSupplier(String model, ModelEvent.ModifyBakingResult event, Map<Item, ResourceLocation> overrides, ModelProperty<ItemStack> property) {
+        ResourceLocation key = ModelResourceLocation.tryParse(model);
+        assert key != null;
+        BakedModel nest = event.getModels().get(key);
+        if (!(event.getModels().get(key) instanceof EggSupplierBakedModel))
+            event.getModels().put(key, new EggSupplierBakedModel(nest, overrides, property));
     }
 
     @SubscribeEvent
-    public static void onBaking(ModelBakeEvent event) {
+    public static void onBaking(ModelEvent.ModifyBakingResult event) {
         //TODO: Allow for collecting custom model resources for each item from somewhere
         Map<Item, ResourceLocation> nest = new HashMap<>();
         nest.put(Items.EGG, DEFAULT_NEST_EGG);
         Map<Item, ResourceLocation> incubator = new HashMap<>();
         incubator.put(Items.EGG, DEFAULT_INCUBATOR_EGG);
         replaceWithEggSupplier("husbandry:nest#", event, nest, NestTileEntity.ITEM_STACK);
-        replaceWithEggSupplier("husbandry:incubator#facing=north", event, incubator, IncubatorTileEntity.ITEM_STACK);
-        replaceWithEggSupplier("husbandry:incubator#facing=east", event, incubator, IncubatorTileEntity.ITEM_STACK);
-        replaceWithEggSupplier("husbandry:incubator#facing=west", event, incubator, IncubatorTileEntity.ITEM_STACK);
-        replaceWithEggSupplier("husbandry:incubator#facing=south", event, incubator, IncubatorTileEntity.ITEM_STACK);
+        replaceWithEggSupplier("husbandry:incubator#facing=north", event, incubator, IncubatorBlockEntity.ITEM_STACK);
+        replaceWithEggSupplier("husbandry:incubator#facing=east", event, incubator, IncubatorBlockEntity.ITEM_STACK);
+        replaceWithEggSupplier("husbandry:incubator#facing=west", event, incubator, IncubatorBlockEntity.ITEM_STACK);
+        replaceWithEggSupplier("husbandry:incubator#facing=south", event, incubator, IncubatorBlockEntity.ITEM_STACK);
     }
 }
