@@ -1,22 +1,22 @@
 package uk.joshiejack.husbandry.client.gui;
 
-import com.mojang.blaze3d.matrix.MatrixStack;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.widget.Widget;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.components.AbstractStringWidget;
+import net.minecraft.network.chat.Component;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.api.distmarker.OnlyIn;
 
 import javax.annotation.Nonnull;
 
 @OnlyIn(Dist.CLIENT)
-public class InformationLabel extends Widget {
-    public InformationLabel(int x, int y) {
-        super(x, y, 120, 24, new TranslationTextComponent("gui.husbandry.noentity"));
+public class InformationLabel extends AbstractStringWidget {
+    public InformationLabel(int x, int y, Font font) {
+        super(x, y, 120, 24, Component.translatable("gui.husbandry.noentity"), font);
     }
 
     @Override
-    public void renderButton(@Nonnull MatrixStack matrix, int mouseX, int mouseY, float partialTicks) {
-        Minecraft.getInstance().font.drawWordWrap(getMessage(), x + 22, y + 8, 120, 4210752);
+    public void renderWidget(@Nonnull GuiGraphics graphics, int mouseX, int mouseY, float partialTicks) {
+        graphics.drawWordWrap(getFont(), getMessage(), getX() + 22, getY() + 8, 120, 4210752);
     }
 }
