@@ -1,10 +1,10 @@
 package uk.joshiejack.husbandry.command;
 
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
-import net.minecraft.command.CommandSource;
-import net.minecraftforge.event.RegisterCommandsEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
+import net.minecraft.commands.CommandSourceStack;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.neoforge.event.RegisterCommandsEvent;
 import uk.joshiejack.husbandry.Husbandry;
 
 @Mod.EventBusSubscriber(modid = Husbandry.MODID)
@@ -12,7 +12,7 @@ public class HusbandryCommands {
     @SubscribeEvent
     public static void registerCommands(RegisterCommandsEvent event) {
         event.getDispatcher().register(
-                LiteralArgumentBuilder.<CommandSource>literal(Husbandry.MODID)
+                LiteralArgumentBuilder.<CommandSourceStack>literal(Husbandry.MODID)
                         .requires(cs -> cs.hasPermission(2))
                         .then(SetHeartsCommand.register())
                         .then(SetHappinessCommand.register())
