@@ -1,20 +1,20 @@
 package uk.joshiejack.husbandry.data;
 
 import joptsimple.internal.Strings;
-import net.minecraft.data.DataGenerator;
-import net.minecraft.item.BlockItem;
-import net.minecraftforge.client.model.generators.ItemModelProvider;
-import net.minecraftforge.client.model.generators.ModelFile;
-import net.minecraftforge.common.data.ExistingFileHelper;
-import net.minecraftforge.fml.RegistryObject;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.data.PackOutput;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.ItemStack;
+import net.neoforged.neoforge.client.model.generators.ItemModelProvider;
+import net.neoforged.neoforge.client.model.generators.ModelFile;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
+import net.neoforged.neoforge.registries.DeferredHolder;
 import uk.joshiejack.husbandry.Husbandry;
-import uk.joshiejack.husbandry.item.HusbandryItems;
-
-import java.util.Objects;
+import uk.joshiejack.husbandry.world.item.HusbandryItems;
 
 public class HusbandryItemModels extends ItemModelProvider {
-    public HusbandryItemModels(DataGenerator generator, ExistingFileHelper existingFileHelper) {
-        super(generator, Husbandry.MODID, existingFileHelper);
+    public HusbandryItemModels(PackOutput output, ExistingFileHelper existingFileHelper) {
+        super(output, Husbandry.MODID, existingFileHelper);
     }
 
     private boolean isFeed(String path) {
@@ -24,9 +24,9 @@ public class HusbandryItemModels extends ItemModelProvider {
     @Override
     protected void registerModels() {
         HusbandryItems.ITEMS.getEntries().stream()
-                .map(RegistryObject::get)
+                .map(DeferredHolder::get)
                 .forEach(item -> {
-                    String path = Objects.requireNonNull(item.getRegistryName()).getPath();
+                    String path = BuiltInRegistries.ITEM.getKey(item).getPath();
                     if (item instanceof BlockItem)
                         getBuilder(path).parent(new ModelFile.UncheckedModelFile(modLoc("block/" + path)));
                     else {
@@ -34,7 +34,7 @@ public class HusbandryItemModels extends ItemModelProvider {
                             withExistingParent(path, mcLoc("item/template_spawn_egg"));
                         } else {
                             String subdir =
-                                    item.getFoodProperties() != null ? "food/"
+                                    item.getFoodProperties(new ItemStack(item), null) != null ? "food/"
                                             : path.contains("treat") ? "treat/"
                                             : isFeed(path) ? "feed/"
                                             : Strings.EMPTY;
